requirements.txt
# Reset to default PyPI for other packages
--index-url https://pypi.org/simple/

# YOLO and Computer Vision
ultralytics==8.3.153
opencv-python==4.11.0.86

# Core scientific computing
numpy==2.3.0
pillow==11.2.1
scipy==1.15.3
pandas==2.3.0
matplotlib==3.10.3

# Machine Learning and Deep Learning
tensorflow>=2.8.0,<2.16.0
keras>=2.8.0

# Utilities and dependencies
pyyaml==6.0.2
requests==2.32.4
tqdm==4.67.1
psutil==7.0.0